+++ driver-observe/src/main/java/fr/ird/driver/observe/dao/data/ps/logbook/SampleSpeciesMeasureDao.java
@@ -28,7 +28,7 @@ public class SampleSpeciesMeasureDao extends AbstractDataDao<SampleSpeciesMeasur
     @Override
     protected void fill(SampleSpeciesMeasure result, ResultSet rs) throws SQLException, ObserveDriverException {
         super.fill(result, rs);
-        result.setSizeClass(rs.getFloat(6));
+        result.setSizeClass(rs.getDouble(6));
         result.setCount(rs.getInt(7));
     }
 }
